requirements.txt
-i https://pypi.org/simple
anyio==3.6.2 ; python_full_version >= '3.6.2'
cachecontrol==0.12.11 ; python_version >= '3.6'
cachetools==5.3.0 ; python_version ~= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3 ; python_version >= '3.7'
cryptography==39.0.1
fastapi==0.92.0
firebase-admin==6.1.0
google-api-core[grpc]==2.11.0 ; platform_python_implementation != 'PyPy'
google-api-python-client==2.78.0 ; python_version >= '3.7'
google-auth==2.16.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2 ; python_version >= '3.7'
google-cloud-firestore==2.9.1 ; platform_python_implementation != 'PyPy'
google-cloud-storage==2.7.0 ; python_version >= '3.7'
google-crc32c==1.5.0 ; python_version >= '3.7'
google-resumable-media==2.4.1 ; python_version >= '3.7'
googleapis-common-protos==1.58.0 ; python_version >= '3.7'
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.14.0 ; python_version >= '3.7'
httplib2==0.21.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
httptools==0.5.0
idna==3.4 ; python_version >= '3.5'
msgpack==1.0.4
proto-plus==1.22.2 ; python_version >= '3.6'
protobuf==4.22.0 ; python_version >= '3.7'
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.5 ; python_version >= '3.7'
pyjwt[crypto]==2.6.0 ; python_version >= '3.7'
pyparsing==3.0.9 ; python_version >= '3.1'
python-dotenv==0.21.1
pyyaml==6.0
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
rsa==4.9 ; python_version >= '3.6'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.25.0 ; python_version >= '3.7'
typing-extensions==4.5.0 ; python_version >= '3.7'
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uvicorn[standard]==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
